refactor: remove the `slot_not_assigned.rs`
